+++ nginx-fastcgi/metadata.rb
@@ -1,10 +1,6 @@
 maintainer       "Alexey Melezhik"
 maintainer_email "melezhik@gmail.com"
 license          "All rights reserved"
-description      "CPAN modules provider (cpan_client provider LWRP)"
+description      "configure nginx to operate with fastcgi application"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.0.9"
-%w{ ubuntu gentoo debian }.each do |os|
-  supports os
-end
-
+version          "0.0.1"